So.. About CTX.
I'll have to ask but my ego is 2 strong to, but I Think CTX is more of a
modifier. That being said, I took it out and i'm testing this method
with self.message.fuckyou for, science bitch. Tbh idk what I'm doing.
Yolo not learning python the reading way.
screw snakes and harry potter.